requirements.txt
openai>=1.0.0
python-dotenv==1.0.1
numpy
faiss-cpu
scikit-learn
fastapi
uvicorn
pandas